remove unused header content in FFI (#396)
