+++ backend/src/tensorflow/compiler/xla/client/client_library.cpp
@@ -16,8 +16,8 @@ limitations under the License.
 #include "tensorflow/compiler/xla/client/client_library.h"
 #include "tensorflow/compiler/xla/client/local_client.h"
 
+#include "local_client.h"
 #include "../../../stream_executor/platform.h"
-#include "client_library.h"
 
 extern "C" {
     LocalClient* ClientLibrary_GetOrCreateLocalClient(
