+++ backend/src/tensorflow/compiler/xla/shape.h
@@ -15,6 +15,4 @@ limitations under the License.
 */
 extern "C" {
     struct Shape;
-
-    void Shape_delete(Shape* s);
 }
