+++ backend/src/tensorflow/core/platform/status.h
@@ -15,6 +15,4 @@ limitations under the License.
 */
 extern "C" {
     struct Status;
-    void Status_delete(Status* status);
-    int Status_ok(Status& status);
 }
