+++ backend/src/tensorflow/stream_executor/BUILD
@@ -18,6 +18,5 @@ cc_library(
     alwayslink = True,
     srcs = glob(["*.cpp"]),
     hdrs = glob(["*.h"]),
-    deps = ["@xla_extension//:xla_extension"],
     visibility = ["//visibility:public"],
 )
